MS48 Custom hook for mutations
